Ya se pueden eliminar clientes sin citas
